some fix.

+++ include/pybind11/internal/object.h
@@ -120,7 +120,7 @@ public:
 
     name(const char* data, int size) : data(py_namev({data, size})) {}
 
-    name(std::string_view str) : name(str.data(), str.size()) {}
+    name(std::string_view str) : name(str.data(), static_cast<int>(str.size())) {}
 
     name(handle h);
 
@@ -198,9 +198,7 @@ public:
     object() = default;
 
     object(const object& other) : handle(other), m_index(other.m_index) {
-        if(other.in_pool()) {
-            object_pool::inc_ref(other);
-        }
+        if(other.in_pool()) { object_pool::inc_ref(other); }
     }
 
     object(object&& other) : handle(other), m_index(other.m_index) {
@@ -210,12 +208,8 @@ public:
 
     object& operator= (const object& other) {
         if(this != &other) {
-            if(in_pool()) {
-                object_pool::dec_ref(*this);
-            }
-            if(other.in_pool()) {
-                object_pool::inc_ref(other);
-            }
+            if(in_pool()) { object_pool::dec_ref(*this); }
+            if(other.in_pool()) { object_pool::inc_ref(other); }
             m_ptr = other.m_ptr;
             m_index = other.m_index;
         }
@@ -224,9 +218,7 @@ public:
 
     object& operator= (object&& other) {
         if(this != &other) {
-            if(in_pool()) {
-                object_pool::dec_ref(*this);
-            }
+            if(in_pool()) { object_pool::dec_ref(*this); }
             m_ptr = other.m_ptr;
             m_index = other.m_index;
             other.m_ptr = nullptr;
@@ -236,9 +228,7 @@ public:
     }
 
     ~object() {
-        if(in_pool()) {
-            object_pool::dec_ref(*this);
-        }
+        if(in_pool()) { object_pool::dec_ref(*this); }
     }
 
     bool is_singleton() const { return m_ptr && m_index == -1; }
